Fix to #1 - use JSON.generate
use JSON.generate rather than .to_json(*a) for json generation. This avoids
a conflict with ActiveSupport core extension, which provides its own to_json, and appears to behave differently.
One important difference that was causing this problem, was that it didn't seem to delegate on to a to_json method within
a nested object.